requirements.txt
opencv-python
numpy
matplotlib
scikit-learn
Pillow
torch>=2.0.0
torchvision>=0.15.0
tqdm
